requirements.txt
python-dotenv
openpyxl
azure-ai-translation-text==1.0.0b1
jieba
flask
azure-cognitiveservices-vision-computervision
pandas
requests
odfpy
validators
